Build/Travis: prevent issues with incompatible PHPCS on PHP < 5.4
YoastCS uses PHP_CodeSniffer 3.x which has a minimum requirement of PHP 5.4.
As the build check running PHPCS is run on PHP 7.2 and devs generally use a reasonably PHP version as well, this shouldn't be a problem.

However, when validating the `composer.json` file during builds, Composer also tests all the scripts defined by the various packages and if this is done on PHP 5.3, the scripts for PHP_CodeSniffer will - predictably - fail.

Removing PHPCS as a dependency for the PHP 5.2/3 builds, together with making sure that the WPSEO `composer install` does not install `dev` dependencies (which would contain PHPCS again), solves this.