Tests: update for changed BrainMonkey has_filter/has_action() behaviour
The WP native `has_filter()` and `has_action()` functions return a boolean if no function name is given, or the priority as an integer if a specific function is being checked or `false` if the function is not attached.

The BrainMonkey mock of the `has_filter()` and `has_action()` functions previously always returned a boolean.

As of version 2.5.0, they will emulate the WP native functions better and return the priority.

This updates the affected tests to check based on the new behaviour of the mock.

Refs:
* https://developer.wordpress.org/reference/functions/has_filter/
* https://developer.wordpress.org/reference/functions/has_action/
* https://github.com/Brain-WP/BrainMonkey/releases